Refactoring ERB rendering and adding an untit test
